gdbstub: Remove a local flag (send_trap) serving an unused purpose of execution flow. (cleanup)
If set to false, it denies entry to the GDBStub::SendTrap function. Only two callers
exist for it: ExceptionRaised() in arm_dynarmic.cpp, and ServeBreak() in armstate.cpp.
However, both callers also call GDBStub::Break() beforehand, which forces 'send_trap'
to true.  Therefore, it looks as if this boolean flag has no real purpose.

After clean-up, it becomes clear that GDBStub::Break is a mere 'is_memory_break' setter.

I'm thinking of moving the RecordBreak and ServeBreak methods (+ underlying bkpt data)
from armstate.cpp over to the GDBStub, so that Break is now RecordBreak, and so that
GDBStub::SendSignal can, without defining a new optional parameter, pull up the address
of a watchpoint hit, (if any) adding it to the "extended stop information" sent to GDB!