gdbstub: Implement BP range check (1/2) in CheckBreakpoint
We test every watchpoint of the given type whose start address is within [0; addr]
where "addr" is the start address of a read/write operation.

** Why the lower bound is 0 inclusive:
In theory, gdb could insert watchpoints of any arbitrary size. Not sure how that's doable in practice in the vanilla gdb client, but I'm playing it safe.

** Why the upper bound is "addr" inclusive:
For this commit's diff alone, we assume all memory operations are 1 byte each - the very same assumption the gdb client makes, given that a stop reply can only specify the address of a watchpoint hit, not the size/width of the memory operation causing it.